staging: comedi: mpc624: change type of private data 'ai_speed'
For aesthetics, change the type of this member to unsigned int.

+++ drivers/staging/comedi/drivers/mpc624.c
@@ -94,7 +94,7 @@ FASTVC-BENCH-CORPUS:linux-main-100k-v1-e0bd41dc-8e12-4f1b-978d-a3645ae59da0
 #define MPC624_SPEED_6_875_HZ	MPC624_OSR(0x1f)
 
 struct mpc624_private {
-	unsigned long int ai_speed;
+	unsigned int ai_speed;
 };
 
 /* -------------------------------------------------------------------------- */
